Refactor create-listing screen by splitting composables accross different files
